requirements.txt
pandas~=2.0.3
scipy~=1.10.1
kornia~=0.7.1
numpy~=1.24.3
torchvision~=0.13.1
imutils~=0.5.4
scikit-image~=0.21.0
scikit-learn~=1.3.2
pillow~=10.2.0
opencv-python==4.9.0.80
openpyxl~=3.1.2